move: content mini dialog hook into components/ui/Content/hooks
